Support incremental compilation
