+++ kson-processor/src/test/kotlin/dev/afanasev/kson/processor/KPropertyKtTest.kt
@@ -1,6 +1,9 @@
 package dev.afanasev.kson.processor
 
-import com.squareup.kotlinpoet.*
+import com.squareup.kotlinpoet.ClassName
+import com.squareup.kotlinpoet.TypeName
+import com.squareup.kotlinpoet.asClassName
+import com.squareup.kotlinpoet.asTypeName
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 
@@ -14,18 +17,17 @@ class KPropertyKtTest {
         val string = ClassName.bestGuess("String")
         assert("string_adapter", string)
 
-        val listOfString = ParameterizedTypeName.get(List::class, String::class)
+        val listOfString = List::class.parameterizedBy(String::class)
         assert("list_string_adapter", listOfString)
 
-        val mapOfStringToInt = ParameterizedTypeName.get(Map::class, String::class, Int::class)
+        val mapOfStringToInt = Map::class.parameterizedBy(String::class, Int::class)
         assert("map_string_int_adapter", mapOfStringToInt)
 
-        val mapOfMap = ParameterizedTypeName.get(Map::class.asClassName(), Int::class.asTypeName(), mapOfStringToInt)
+        val mapOfMap = Map::class.asClassName().parameterizedBy(Int::class.asTypeName(), mapOfStringToInt)
         assert("map_int_map_string_int_adapter", mapOfMap)
     }
 
     private fun assert(expectedName: String, type: TypeName) {
         assertThat(typeToAdapterName(type)).isEqualTo(expectedName)
     }
-
 }
